Modified corrections
