+++ src/app/layout.tsx
@@ -15,7 +15,6 @@ export const metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
-console.log(domain, client_id);
 
 export default function RootLayout({
   children,
