--- src/app/login/page.tsx
@@ -1,9 +1,0 @@
-"use client"
-
-import LoginComponent from '../../components/logIn'
-
-const login = () => {
-    return <LoginComponent/>
-}
-
-export default login
