--- src/app/signUp/page.tsx
@@ -1,9 +1,0 @@
-"use client"
-
-import SignUpComponent from "@/components/signUp"
-
-const signUp = () => {
-    return <SignUpComponent />
-}
-
-export default signUp
