+++ src/components/productAddModal/index.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import { useState } from 'react';
 import { IProductAddModal } from './interface';
+import MinusIcon from '../../assets/svg/Minus Icon.svg';
+import PlusIcon from '../../assets/svg/Plus Icon.svg';
+import Image from 'next/image';
 
 const ProductAddModal: React.FC<IProductAddModal> = ({
 	setIsModalShown,
@@ -52,29 +55,7 @@ const ProductAddModal: React.FC<IProductAddModal> = ({
 								id='modal-text'
 								className='mt-2 flex flex-row w-full justify-evenly'>
 								<button onClick={decrementCount}>
-									<svg
-										height='24px'
-										width='24px'
-										id='Capa_1'
-										viewBox='0 0 44.143 44.143'
-										fill='#000000'>
-										<g
-											id='SVGRepo_bgCarrier'
-											stroke-width='0'></g>
-										<g
-											id='SVGRepo_tracerCarrier'
-											stroke-linecap='round'
-											stroke-linejoin='round'></g>
-										<g id='SVGRepo_iconCarrier'>
-											<g>
-												<g>
-													<path
-														fill='#030104'
-														d='M34.738,21.073H9.404c-0.553,0-1,0.447-1,1s0.447,1,1,1h25.334c0.553,0,1-0.447,1-1 C35.738,21.519,35.291,21.073,34.738,21.073z'></path>
-												</g>
-											</g>
-										</g>
-									</svg>
+									<Image className='w-10 h-10' src={MinusIcon} alt='MinusIcon'></Image>
 								</button>
 								<div>
 									<input
@@ -87,26 +68,7 @@ const ProductAddModal: React.FC<IProductAddModal> = ({
 									<p className='pt-2'>${count * price}</p>
 								</div>
 								<button onClick={incrementCount}>
-									<svg
-										fill='#000000'
-										width='40px'
-										height='40px'
-										viewBox='-3.2 -3.2 38.40 38.40'
-										transform='rotate(0)'>
-										<g
-											id='SVGRepo_bgCarrier'
-											stroke-width='0'></g>
-										<g
-											id='SVGRepo_tracerCarrier'
-											stroke-linecap='round'
-											stroke-linejoin='round'
-											stroke='#CCCCCC'
-											stroke-width='0.8960000000000001'></g>
-										<g id='SVGRepo_iconCarrier'>
-											{' '}
-											<path d='M9,17h6v6a1,1,0,0,0,2,0V17h6a1,1,0,0,0,0-2H17V9a1,1,0,0,0-2,0v6H9a1,1,0,0,0,0,2Z'></path>{' '}
-										</g>
-									</svg>
+									<Image className='w-10 h-10' src={PlusIcon} alt='PlusIcon'></Image>
 								</button>
 							</div>
 						</div>
