+++ src/redux/services/category/inteface.ts
@@ -1,0 +1,7 @@
+export interface IApiResponseCategories {
+    status: string
+    payload: [{
+        name: string,
+        image: string
+    }]
+}
