+++ src/redux/services/user/interface.ts
@@ -1,0 +1,17 @@
+export interface Iuser {
+	lastName: string;
+	firstName: string;
+	email: string;
+	image: string;
+	cellphone: number;
+}
+
+export interface IpostResponse {
+	id: string,
+	user: Iuser
+}
+
+export interface IApiUsersResponse {
+	status: string;
+	payload: Iuser[] | string;
+}
